+++ b2/_internal/_cli/argcompleters.py
@@ -16,9 +16,11 @@ from itertools import islice
 
 
 def bucket_name_completer(prefix, parsed_args, **kwargs):
+    from b2sdk.v2 import unprintable_to_hex
+
     from b2._internal._cli.b2api import _get_b2api_for_profile
     api = _get_b2api_for_profile(getattr(parsed_args, 'profile', None))
-    res = [bucket.name for bucket in api.list_buckets(use_cache=True)]
+    res = [unprintable_to_hex(bucket.name) for bucket in api.list_buckets(use_cache=True)]
     return res
 
 
@@ -28,7 +30,7 @@ def file_name_completer(prefix, parsed_args, **kwargs):
 
     To limit delay & cost only lists files returned from by single call to b2_list_file_names
     """
-    from b2sdk.v2 import LIST_FILE_NAMES_MAX_LIMIT
+    from b2sdk.v2 import LIST_FILE_NAMES_MAX_LIMIT, unprintable_to_hex
 
     from b2._internal._cli.b2api import _get_b2api_for_profile
 
@@ -41,7 +43,7 @@ def file_name_completer(prefix, parsed_args, **kwargs):
         fetch_count=LIST_FILE_NAMES_MAX_LIMIT,
     )
     return [
-        folder_name or file_version.file_name
+        unprintable_to_hex(folder_name or file_version.file_name)
         for file_version, folder_name in islice(file_versions, LIST_FILE_NAMES_MAX_LIMIT)
     ]
 
@@ -50,7 +52,7 @@ def b2uri_file_completer(prefix: str, parsed_args, **kwargs):
     """
     Complete B2 URI pointing to a file-like object in a bucket.
     """
-    from b2sdk.v2 import LIST_FILE_NAMES_MAX_LIMIT
+    from b2sdk.v2 import LIST_FILE_NAMES_MAX_LIMIT, unprintable_to_hex
 
     from b2._internal._cli.b2api import _get_b2api_for_profile
     from b2._internal._utils.python_compat import removeprefix
@@ -60,7 +62,10 @@ def b2uri_file_completer(prefix: str, parsed_args, **kwargs):
     if prefix.startswith('b2://'):
         prefix_without_scheme = removeprefix(prefix, 'b2://')
         if '/' not in prefix_without_scheme:
-            return [f"b2://{bucket.name}/" for bucket in api.list_buckets(use_cache=True)]
+            return [
+                f"b2://{unprintable_to_hex(bucket.name)}/"
+                for bucket in api.list_buckets(use_cache=True)
+            ]
 
         b2_uri = parse_b2_uri(prefix)
         bucket = api.get_bucket_by_name(b2_uri.bucket_name)
@@ -72,7 +77,7 @@ def b2uri_file_completer(prefix: str, parsed_args, **kwargs):
             with_wildcard=True,
         )
         return [
-            f"b2://{bucket.name}/{file_version.file_name}"
+            unprintable_to_hex(f"b2://{bucket.name}/{file_version.file_name}")
             for file_version, folder_name in islice(file_versions, LIST_FILE_NAMES_MAX_LIMIT)
             if file_version
         ]