+++ b2/_internal/_cli/shell.py
@@ -10,6 +10,7 @@
 
 import os
 import os.path
+import shutil
 from typing import Optional
 
 
@@ -19,3 +20,15 @@ def detect_shell() -> Optional[str]:
     if shell_var:
         return os.path.basename(shell_var)
     return None
+
+
+def resolve_short_call_name(binary_path: str) -> str:
+    """
+    Resolve the short name of the binary.
+
+    If binary is in PATH, return only basename, otherwise return a full path.
+    This method is to be used with sys.argv[0] to resolve handy name for the user instead of full path.
+    """
+    if shutil.which(binary_path) == binary_path:
+        return os.path.basename(binary_path)
+    return binary_path