+++ doc/source/conf.py
@@ -42,7 +42,6 @@ from b2._internal.version_listing import LATEST_STABLE_VERSION
 
 B2 = importlib.import_module(f'b2._internal.{LATEST_STABLE_VERSION}.registry').B2
 
-
 # -- General configuration ------------------------------------------------
 
 # If your documentation needs a minimal Sphinx version, state it here.
@@ -53,12 +52,8 @@ B2 = importlib.import_module(f'b2._internal.{LATEST_STABLE_VERSION}.registry').B
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.intersphinx',
-    'sphinx.ext.ifconfig',
-    'sphinx.ext.viewcode',
-    'sphinx.ext.coverage',
-    'sphinxarg.ext'
+    'sphinx.ext.autodoc', 'sphinx.ext.intersphinx', 'sphinx.ext.ifconfig', 'sphinx.ext.viewcode',
+    'sphinx.ext.coverage', 'sphinxarg.ext'
 ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -74,10 +69,10 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'B2_Command_Line_Tool'
+project = 'B2_Command_Line_Tool'
 
 year = datetime.date.today().strftime("%Y")
-author = u'Backblaze'
+author = 'Backblaze'
 copyright = f'{year}, {author}'
 
 # The version info for the project you're documenting, acts as replacement for
@@ -181,8 +176,8 @@ latex_elements = {
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
     (
-        master_doc, 'B2_Command_Line_Tool.tex', u'B2\\_Command\\_Line\\_Tool Documentation',
-        u'Backblaze', 'manual'
+        master_doc, 'B2_Command_Line_Tool.tex', 'B2\\_Command\\_Line\\_Tool Documentation',
+        'Backblaze', 'manual'
     ),
 ]
 
@@ -191,7 +186,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'b2_command_line_tool', u'B2_Command_Line_Tool Documentation', [author], 1)
+    (master_doc, 'b2_command_line_tool', 'B2_Command_Line_Tool Documentation', [author], 1)
 ]
 
 # -- Options for Texinfo output -------------------------------------------
@@ -201,14 +196,13 @@ man_pages = [
 #  dir menu entry, description, category)
 texinfo_documents = [
     (
-        master_doc, 'B2_Command_Line_Tool', u'B2_Command_Line_Tool Documentation', author,
+        master_doc, 'B2_Command_Line_Tool', 'B2_Command_Line_Tool Documentation', author,
         'B2_Command_Line_Tool', 'One line description of project.', 'Miscellaneous'
     ),
 ]
 
 # Example configuration for intersphinx: refer to the Python standard library.
 intersphinx_mapping = {'python3': ('https://docs.python.org/3', None)}
-
 
 white_spaces_start = re.compile(r'^\s*')
 
@@ -219,12 +213,12 @@ def setup(_):
     reasonable way to piggy back that behaviour. Checking if the new file contents would the same as the old one
     (if any) is important, so that the automatic file-watcher/doc-builder doesn't fall into an endless loop.
     """
-    main_help_text = str(B2.lazy_get_description())
+    main_help_text = str(B2.lazy_get_description(NAME='b2'))
     main_help_text = textwrap.dedent(main_help_text)
 
     main_help_path = path.join(path.dirname(__file__), 'main_help.rst')
     if path.exists(main_help_path):
-        with open(main_help_path, 'r') as main_help_file:
+        with open(main_help_path) as main_help_file:
             if main_help_file.read() == main_help_text:
                 return
     with open(main_help_path, 'w') as main_help_file: